fix интерфейсов и создания бинов
